Update BLadd16204.xml
Budge includes a negative facsimile of ff. 1-25 in his book. As he notes in the introduction and as one can see in the facsimile, after the ጸሎት፡ ዘመንገደ፡ ሰማይ፡, a part of ልፋፈ፡ ጽድቅ፡, but before the ትርጓሜ፡ ፊደል፡ (? this is absent in the facsimile) is መጽሐፈ፡ አርድእት፡. There are three entries for this text in BM, but I can't tell which is which since they don't have the text, so I went with the most generic.