Update BDLaethc14.xml (#1345)
* Update BDLaethc14.xml

* Update BDLaethc14.xml

* Art descs for c 14, d20, e 0

Amended on the basis of Thea's comments. I note that I am only using the "term" function for details I think might be indexed. For index, in cases such as e.g. "ሉቃስ።" should we be using per name to ensure these will be indexed or is there a way to tell the app to pick these up?

* Update BDLaethc14.xml

* Update BDLaethc14.xml

* Update BDLaethd20.xml

* Update BDLaethe30.xml

Co-authored-by: thea-m <thea.m@hotmail.de>
Co-authored-by: thea-m <dorothea.reule@gmail.com>