Update BMLsmarco741.xml
I added physDesc and additions and checked everything in the given images. I have seen that scan no. 55 and 56 are identical (showing folia 51v-52r).
I was not able to identify another hand on f. 36r-v, as Marrassini did. I am not sure, whether I should follow Marrassini in the record.
I was not able to find Andreas Datius Florentinus, who is mentioned in an addition (e3) on f. 1v.