Updated p5 and p6 in BMLacq306
The unidentified text in p5_i3 has four words in common (ኂሩትከ፡ አመፍቀሬ፡ ሰብእ፡ በእንተ፡) with the second prayer of LIT3948MashafaKrS. Is it enough to state that supposition or does it need more research? BSBaeth33 is a witness, but images from this manuscript are not available to me.
For p6_i2 (Maṣḥafa nuzaze) Marrassini quoted an additio written in the margins as incipit. I quoted this in a note and took the text in the colums as incipit. Apparently the other text was inserted in a later time. Is it necessary to quote this insertion one more time as addition later on?