Update BMLacq810.xml
The encoding of BMLacq810 took me long time. The manuscript is messed up and the folios are not in correct order. My assumption is, that many folios are missing. Apparently the order of the first folios was different, when Marrassini described the manuscript because the incipits several times do not match to the given folios. I kept the order and foliation in the manuscript. I could not identify many prayers and textparts. All manuscripts, that I checked, have another order and do not have IDs for the individual prayers as well. Marrassini frequently refers to the manuscripts Uppsala 18, Vatican 74, Vatican 115 and others. I did not check these accounts, because the images in all of these cases are not available. I noted the references in the way as Marrassine did, but I am not sure whether the numbers refer to folios or to numbers for text parts in the catalogues. Is it neccessary to enhance the description by creating an issue for each of the not yet identified textual units? I rather ask in advance, because they are very many. I will create an ID for the owner and will insert the ID in a minute.