Update date BMLacq810
The date of the manuscript can be given more precisely thanks to the identification of the scribe by Denis, who identified the same scribe in many other manuscripts that are precisely dated by colophons and other indications. Is it useful or necessary to give the reason for the distinct date in the manuscript record?