Created BNFet337 and updated content of BMLacq810
Created a record for BNFet337 and inserted the reference to the record of BMLacq810. Unfortunatly the oxygen did not validate my new record responding: element "TEI" incomplete; expected element "facsimile", "sourceDoc" or "text"
I copied the two records of JRyl7 and BNFet337 from BMLacq306 and BMLacq810 where this problem does not occur. I don't know what was going wrong and how to fix it.