Update BMLacq810.xml (#2049)
* Update BMLacq810.xml

The encoding of BMLacq810 took me long time. The manuscript is messed up and the folios are not in correct order. My assumption is, that many folios are missing. Apparently the order of the first folios was different, when Marrassini described the manuscript because the incipits several times do not match to the given folios. I kept the order and foliation in the manuscript. I could not identify many prayers and textparts. All manuscripts, that I checked, have another order and do not have IDs for the individual prayers as well. Marrassini frequently refers to the manuscripts Uppsala 18, Vatican 74, Vatican 115 and others. I did not check these accounts, because the images in all of these cases are not available. I noted the references in the way as Marrassine did, but I am not sure whether the numbers refer to folios or to numbers for text parts in the catalogues. Is it neccessary to enhance the description by creating an issue for each of the not yet identified textual units? I rather ask in advance, because they are very many. I will create an ID for the owner and will insert the ID in a minute.

* Update BMLacq810.xml

I inserted the newly created ID here. Is it necessary or useful to mention the owner in <origin> as well?

* Update BMLacq810

According to Denis, this manuscript might be from the 17th or 18th century.

* Update date BMLacq810

The date of the manuscript can be given more precisely thanks to the identification of the scribe by Denis, who identified the same scribe in many other manuscripts that are precisely dated by colophons and other indications. Is it useful or necessary to give the reason for the distinct date in the manuscript record?

* Create UppEt18

Created stub for Uppsala 18

* Updated BMLacq810

I made the recommended changes and created stubs for ANLcr71 and ONBAeth18 in order to refer to them. I did create stubs for JRyl and BN(Gr.) yet, because I didn't have the catalogues by Strelcyn at my disposal.

* Update BMLacq810.xml

I created a note on the provenance of BMLacq810

* Update BMLacq810.xml

Updated provenance according to Denis note in BMLacq306.

* Updated BMLacq810 and created record for JRyl7

I created a new record for JRyl7 and inserted this reference to the BMLacq810 record.

* Created BNFet337 and updated content of BMLacq810

Created a record for BNFet337 and inserted the reference to the record of BMLacq810. Unfortunatly the oxygen did not validate my new record responding: element "TEI" incomplete; expected element "facsimile", "sourceDoc" or "text"
I copied the two records of JRyl7 and BNFet337 from BMLacq306 and BMLacq810 where this problem does not occur. I don't know what was going wrong and how to fix it.