Encoded BMLacq720
Again the pagination is problematic, because the number 63 was given twice to folio 63 and 64.
I created new entries for the persons mentioned.
I was not sure, whether it was necessary to encode the dates of commemoration (21st Miyazya, 21st Maggabit and 21st Sane) and whether I did this appropriately and detailed enough.
In seeking for the correct LIT, I came across EMML2133 and updated this record with the same narratives included.