Update BDLaethe38.xml
Longer deco note @MassimoVilla please check my captions, I am uncertain about some transcriptions. Also, you might want to check the palaeography of EMML 2349 for comparative purposes, up to you of course!

Do you think the handwriting of the captions of the miniatures and that of the text of this manuscript suggest they were written by the same hand?

Look forward to feedback!