Updated BMLacq776
Because Marrassini gave a very detailed description of the contents of this manuscript, I wanted to be precise as well. Unfortunately, I was not able to identify the works or individual textparts of LIT1295 in some cases:
ms_i1.4.2
ms_i1.4.3 (described as "Michael sent to Samson" by Marrassini)
ms_1.5.2
ms_1.11.2
ms_1.12.2

It is likely that in fact these parts can be found in other witnesses of Homily of Mikāʾel, but I could not check in manuscripts but relied on the quoted incipits for BAVet82 and EMIP00016, because I had no access to images for these two or for other related manuscripts.
I do not understand which manuscript was indicated by Marrassini with A 84 = N 137 in ms.i1.4.1. It is not mentioned in the introductory text. If this note is not very important, it can also deleted. Also I am not sure, whether I encoded the references to manuscripts correctly.
In describing the quire structure, I differ from Marrassini in quire 10. His account does not match the total number of folios.
Even more important is another difference in the description of hands. Marrassini thought, there are at least five hands visible, but according to my opinion, there are only two. For the larger part of the manuscript, the size of the pen changes frequently as well as the saturation of the ink, but special features of hand 1 remains constant throughout the codex. It would be nice, if somebody else may have a look to it and may confirm or falsify my judgement!
Marrassini also gives a lot of general information on other manuscripts, that I did not record.
A number of new IDs for works and for persons, related to this manuscript will follow in a minute.