Update BMLacq680
The images of BMLacq680 are not complete. The scan images for folio 6v-7r (after scan page 8) as well as for folio 179v-180r (after scan 180). It can be discerned from the folio marks (bottom left) and from the text, that is not fitting.

In encoding the quire structure and the description of hands, I differed a lot from P. Marrassini. As well the last textpart (ms_i3 and ms_i4) are encoded independently contrary to Marrassinis account. Also I treated the index on f. 4v, I as an addition because it is by another hand as most of ms_i1 but not as the incipit of ms_i1, as Marrassini did.