Update BMLacq680
I updated a couple of PRS-IDs and LOC-IDs, that i have created or figured out lately. I also added the previous owner in a note (e7), that was probably famous empress Zawditu. I remember, there was some other way of encoding ownership notes, but I do not remember quiet well, what was the preferred one.
I think, I still need to encode a note on folio 6v-7r and 179v-180r, that is missing among the scan images, but must be exist in the manuscript itself. How can I mention this? Or may these images to be found?