Updated BMLor148
I updated BMLor148. Again the description of quires and the description of hands is different from Marrassinis descriptions.
For the description of ms_i3 (Senodos) I took BLorient481 and BLorient794 as examples and for this reason encoded the Indices of some canons seperately, although I think, that this is not the perfect way of encoding it.
The astronomical treatises in the back part of the manuscript (ms_i5, ms_i6, ms_i7, ms_i8, ms_i10) I did only a minimal record and left a more precise description to @DenisNosnitsin1970, who asked me to leave this to him, since he is dealing with a number of calendaric works in manuscript at the moment.
As Marrassini, I could not identify the moral teaching in ms_i9.