Created BMLacq38 (#2074)
* Created BMLacq38

I created a record for BMLacq38.
I defined the quire structure different from Marrassini, because I think he was mistaken.
I was not able to identify several titles as works. Some of the chronologies and computus may not have an independant circulation. About ms_i2 and ms_i3 there are issues still open (#2072 #2073). It is yet to be discussed, whether there are work-ID necessary for these. Do you think, I should also open issues for the other ms-Items?
About the dating taken from ms_i9, I am quite confused. Marrassini is convinced, that it must be the 23rd year of the reign of Lebna Dengel (for a reason, I don't know) and the 183rd year of grace (ዓመተ፡ ምሕረት፡), saying that both is 1530 in our calender. But year of grace (ዓመተ፡ ምሕረት፡) as far as I know is counted according to the martyrs of Diocletian in the 4th century. 183 AM should be 542, what is an impossible date for this manuscript. Can anybody help me out here?

* Update BMLacq810.xml

* Updated BMLacq38

I updated BMLacq38 according to various comments you made. I think I adressed every comment and issue except for LIT2455Tergwa (Issue #2335), that I will check later on, when I have obtained the edition of Cowley or the manuscript BritLib Orient. 13830.
I updated also some other records concerning the encoding of the quire structure.

* Update BMLacq38

Updated hand note and title ms_i3.

* Update BMLacq38

I updated the title-ref for ms_i1 and deleted the note, because the information about EMML 1389 as the only additional witness is now encoded to the Work-ID.

* Update BMLacq38

I updated ms_i2 as discussed in issue #2335.

---------

Co-authored-by: thea-m <thea.m@hotmail.de>