Tanasee1
I created a physDesc and a ContentDesc for Tanasee1 according to the catalogue of Hammerschmidt. I found, that Hammerschmidt is not correct on the number of folios, that is 239 in fact and the folio numbers for the folios more that 200 are given with one to high as I indicated in my foliation description. I encoded all textparts and additions according to the true number of folios. It would be helpful to include the new true folio numbers manually in the images. I can do that, if you send me the images and tell me, how I can do it.
Unfortunately the manuscript is available only in microfilm images, unless better pictures must exist, because the decorations are reproduced in coloured pictures and one article by Roger Schneider has a better image from a folio. It would be nice to acquire better pictures or get new ones, but that might be not very easy, as discussed in issue #2416. With some exceptions, the text is well readable anyways, but the quire structure is not discernable and not given in Hammerschmidts catalogue. For this reason, we can not be certain, if the last quire with many additions was added later to the codex in a rebinding, as I hypothesized in my collation description and in history.
I encoded very many personal names and also some place names. Some of them have been already discussed in issue #2410, #2410, #2421, #2429, #2434 and #2434.
For the encoding of decorations, I have opened #2428. I include Guesh in this pull request, so he can also have an eye on the outcome.