Update BMLacq680
I updated BMLacq680 according to P. Marrassinis catalogue. Once again, I needed to give a quire structure that is consibarably different to Marrassinis description.
I found foliation marks, that Marrassini did not mention. I kept the foliation as written in the manuscript for the record unless it is contrary to Marrassini description, who described the first leaves as protection leaves with I-IV. However in describing hands and contents, Marrassini is not consisting in his folio numbers, counting the protection leaves from time to time and sometimes not.
In the description of hands, I found as well many different hands, that Marrassini did not count and other round, I found various folio to be written by the same hand that Marrassini considered to be two or more.
I decided to give the index to the theological treatises in ms_i1 as a div. I do not know, whether it is necessary to delete strings that refer to the transscription of transkribus in turn.
I had a short glance in the transkribus files and I mentioned, that the folio numbering is following Marrassinis system and does not count the protective leaves. The text of ms_i1.2 is given with folio 1r and not with 5r as it is in my record. Is that a problem and how can it be solved eventually?