Update BLorient501 (#2187)
* Update BLorient501

Ted Erho asked me to produce a contribution on BLorient501 for the Handbook of Biblical Manuscripts. I checked the data from the catalogue of Wright and added some additional information. Most of them are of minor significance.
However, I am puzzled about the additions on the last folios and whether they may be magical in content and if I can encode them as magical texts or rather as unclear. I understand hardly any word in a1 and a2.
I added some title to the secondary bibliography, that Ted recommended to me. Alas, I was not able to obtain Perrone/Norelli 1995 and Löfgren 1930. For this reason, I could not check it and could not give a cited range of pages.

* Update BLorient501

Updated record according to Dorotheas requests.

* Update BLorient501

Updated additions and extras according to Denis' and Dorothea's comments.

* Update BLorient501

Updated decoNotes, OwnershipNote and bibliographical information on Löfgren according to Eugenias comments.

* Update BLorient501

Again some minor corrections that I came across.