Update BMLacq680
This is the record, that I mistakenly encoded for a second time. I have compared the two records and found nothing, that needs to be revised in this record, which is indeed better discussed and worked out. In the description of the hands some differences came to the surface, but I refrain from analysing the numerous hands again in detail, because it would take too much time and would bring little progress. Instead, I suggest closing and deleting the other branch and merging this one when everything is settled and OK. I have only made some minor changes to this May/June record.