Update BMLacq679
I updated BMLacq679 by adding four newly created PRS ID in a1.
I created four PRS IDs for people mentioned in BMLacq679 as discussed in #2310.

I came to the conclusion that the person PRS14220HeggaYohannes is the father of PRS14221Menilek.

I am still puzzled as to whether it is possible to establish a family relationship for the others as well.
PRS14222TawaldaMadhen can be seen to be the heir or heiress of the former, if we accept this name in connection with አርስት (succession, inheritance, etc.).

PRS14223Maqaryos is mentioned as ራብአዊ, but it is not clear what ራብአዊ stands for. He could be the fourth in the line of succession (e.g. the great-grandson), or the fourth in the list of blessed people. It is unlikely that he is a high-ranking dignitary such as a patriarch or abbot, as he has no title to indicate this.

The text is given twice on two successive folios. There is some uncertainty about the name Həgga (ʾƏgziʾabəḥer) Yoḥannəs, as he is mentioned once as Həgga Yoḥannəs and the second time as Həgga ʾƏgziʾabəḥer Yoḥannəs. However, I think that the encircling of ʾƏgziʾabəḥer tells us that the name ʾƏgziʾabəḥer was a mistake of the scribe.

<img width="333" alt="Screenshot 2023_Maqaryos" src="https://github.com/BetaMasaheft/Persons/assets/40291787/7e788751-d953-40c9-9e5f-76fed47e960b">

NB. I did not understand how to reference an issue. When I press the 'reference' button in the taskbar, nothing changes except for the # that is introduced.