BMLacq784
I updated the record of BMLacq784 according to the catalogue of Marrassini.
In ms_i2 I encoded several ethnic names as ሸንቀላ, አገው, አምሀራ, ትግራይ as ethnic names or place names. Even though, this is a magical text, I think, they are used as ethnical names rather than as names for demons. I would like to create new IDs also for ትርኵ and ቅማንት, which do not exist, yet, according to my exploration.