Update BMLacq680 (#2131)
* Update BMLacq680

The images of BMLacq680 are not complete. The scan images for folio 6v-7r (after scan page 8) as well as for folio 179v-180r (after scan 180). It can be discerned from the folio marks (bottom left) and from the text, that is not fitting.

In encoding the quire structure and the description of hands, I differed a lot from P. Marrassini. As well the last textpart (ms_i3 and ms_i4) are encoded independently contrary to Marrassinis account. Also I treated the index on f. 4v, I as an addition because it is by another hand as most of ms_i1 but not as the incipit of ms_i1, as Marrassini did.

* Update BMLacq680

I updated a couple of PRS-IDs and LOC-IDs, that i have created or figured out lately. I also added the previous owner in a note (e7), that was probably famous empress Zawditu. I remember, there was some other way of encoding ownership notes, but I do not remember quiet well, what was the preferred one.
I think, I still need to encode a note on folio 6v-7r and 179v-180r, that is missing among the scan images, but must be exist in the manuscript itself. How can I mention this? Or may these images to be found?

* Updated BMLacq680

I updated two work ID: LIT4035SenkessarDL as recommended by Thea; LIT6833Salam as discussed in Issue #2367.

* Update physDesc BMLacq680

A added a note to the physDesc, that two openings are missing in the image viewer, most likely, because they have not been digitized. See Issue #2369.

* Update BMLacq680

This is the record, that I mistakenly encoded for a second time. I have compared the two records and found nothing, that needs to be revised in this record, which is indeed better discussed and worked out. In the description of the hands some differences came to the surface, but I refrain from analysing the numerous hands again in detail, because it would take too much time and would bring little progress. Instead, I suggest closing and deleting the other branch and merging this one when everything is settled and OK. I have only made some minor changes to this May/June record.

* Updated idno facs in BMLacq680

I changed idno facs to 309 according to Eugenias request.

* Update BMLacq680

I changed all ə for ǝ, as Eugenia has requested it.