Update BMLacq821 (#2224)
* Update BMLacq821

I updated the record for BMLacq821. To my surprise, the transcription by transcribus was inserted in the ms-file. It may be necessary to split this record into two to have the transcribus record seperately.

* Update BMLacq821

I updated the record by adding a general ID (LIT5888AsmatPrayer) for the protective prayer in msContents.