Update BMLacq384 (#2199)
* Update BMLacq384

I updated the record on BMLacq384 according to the catalogue description of Marrassini. Again, there is one image missing in the viewer. It is the opening 6v-7r. It is clear from Marrassini description of the quire structure that is in accordance with my own observation (quadernion with 8 leaves and no visible stubs) but also from the text: 6r is ending with ኀበ፤ but the next folio on image (no. 13 in the viewer) starts with ለ፡ ይስሕቱ፡) that do not fit together. For this reason, it might be of advantage to introduce folio number manually. There are some pencilled folio number in the manuscript, but only on a very few folios.

* Changed idno facs in BMLacq384

I changed the idno facs according to Eugenias request.

* Update BMLacq384

I changed all ə for ǝ, as Eugenia has requested it.

* Update BMLacq384

I updated the record according to Dorotheas request.

* Update BMLacq384

I updated the record according to Dorotheas suggestion concerning the use of LIT5020Hemamatihu for all the recensions of this work as proposed in https://github.com/BetaMasaheft/Documentation/issues/2445#issuecomment-1842690250