BMLor403 (#2055)
* BMLor403

I recorded the BMLor403. In the question of hands, I disagree with P. Marrassini who defined two hands for h1. I am also not sure about the dating of palaeographic grounds, but I kept what Marrassini stated, although 15th century appears very early to me.
I found the owner erased many times throughout the manuscript. I wanted to mention this person in a particDesc, but I was not able to figure out, where to insert such a particDesc. For that reason I mention him in every msItem, in which he occurs (or not occurs any more) with all its locus precisely. Can we keep it like this or do I need to make a particDesc in addition? I will create a PRS-ID for this person in a minute, because it appears to me hardly possible, that this person with the name Səmʿon can be reasonably identified with any existing PRS-ID.
En passant I also updated BNFet58 in one detail, because I find the certainty of identifying one particular msItem could be raised.

* Update BMLor403

Updated contents and ownership note.

* Update BNFet58.xml

* Update BMLor403.xml

* Update BMLor403.xml

Included dating method in <origDate>.

* Updated Dating BMLacq306 and BMLor403

I checked all the records, that I have encoded so far and found the problem of using "datingmethod" instead of <lettering> in BMLacq306. I think I cured it now.

* Update BMLor403.xml

* Update BMLor403.xml

* Update BMLacq306 and BNFet58

I updated the records according to Denis requests.

---------

Co-authored-by: thea-m <thea.m@hotmail.de>
Co-authored-by: Eugenia <eugenia.sokolinski@uni-hamburg.de>