Create BLorient12510
I created a record for BLorient12510 based on the description in Strelcyn's catalogue but without access to images.
I encoded the parts of the Gospel that are marked in the manuscript to be read on each day of the week as ms_item, as it is described in the catalogue.