Update Frankfurt mss. on Walda Amid
I updated the records of FSUor134 and FSUor40. The principal change is, that I exchanged the general record LIT4723TarikaWaldaAmid for the more specific LIT6936TarWaAmidBeta in FSUor134 as it is necessary due to the latest reworking of the work records on the Tarika Walda Amid. However, I did some minor changes in both records as well, i.e. I deleted the note on the non-existence of Zarʾa Yāʿqob in the Translation note. This note can only be understood from the point of view of one single recension, as Goldschmidt assumed. Nowadays, it is clear, that there have been two translations and that the second translation was done in the time of Lebna Dengel. For that reason it is nothing extraordinary in this translation note in comparison to other witnesses of the beta recension.
In FSUor40, I found nothing, that needs to be changed. I just made a minor change on the alledged participation of the prominent Liq ʾAṣqu. The fact, that Liq ʾAṣqu  copied this codex by his own hand is reported in Rüppells Reisebericht, which is not always reliable, but to my knowledge Liq ʾAṣqus name does not occur in the manuscript.