Update FSUor40
I updated the record by adding a note concerning the hand of Liq ʾAṣqu. I was puzzled, how to classify this note. It is in fact a note concerning the scribe, but it is not from the alledged scribe himself, but afterwards by Rüppell. For this reason, I classified it as a MixedNote, rather than as a ScribalNote.