Create BLorient13265
I created a record according to the catalogue of S. Strelcyn. I struggled with LIT5066MiraclesMichael in ms_i5. I could not identify the exact miracle, but I can confirm on the basis of comparison of the incipits, that it is not among the twelve miracles that are recorded as LIT4155MiraclesMichael (even though the miracle on the poor man became rich seems to have a similar content). This is why, I decided to use LIT5066MiraclesMichael for this record.