Create BLorient12467 (feb24)
I created a record for BLorient12467 according to the description of S. Strelcyn as a composite manuscript. For p1_i1 I figured out a similarity to LIT1645Homily having in common to በእንተ፡ ፀአተ፡ ነፍስ፡ . I think this correlation is to weak to make an identification. For this reason, I put this hypothesis in a note.
In BLorient12510, I deleted the string incipit, that was included twice in the record.