Create BLorient6466 (CH_march24)
I created a record for BLorient6466. I decided to make up the record as for a composite manuscript. p1, p2 and p4 are certainly nothing more than flyleafs that have been bound to the codex in order to protect the psalter as the main part (p3). However, several distinct information about these flyleaves (sive, provenance as a magic scroll) is given in Strelcyn's description, so I wanted to include this into the record.
I could not identify the prayer in p2_i1 as brougt up in https://github.com/BetaMasaheft/Documentation/issues/2497 recently and I suggest to create a new LIT ID for it.