Update BLorient12467
I updated the subdivision and title markup for LIT1828Mahale, LIT2362Songof and LIT2509Weddas.
I intended to do the same for BLorient6466, but that record was recently merged and is not yet in the master.
Personally, I am not convinced that marking LIT2362Songof should be done in this way without consulting the manuscripts and edition. I wonder why Strelcyn has divided the text into five sections, but for each section gives a number such as (I, I - II, 7) etc.
If the 'sections' in Strelcyn's description are 'canticles' according to our Bm definition, what do the Roman numbers in brackets refer to? Presumably the numbers in brackets refer to what is described as 'canticles' in the edition (and in the BM), and the 'sections' are subdivisions according to the prescribed reading in the service. Do you think I should leave it as it is, or do you think it would be better to return to the more careful encoding and describe the subdivisions as 'sections', as Strelcyn did?