Created BLorient4930 (CH_march24)
I created a record for BLorient4930 according to Strelcyns catalogue.

For ms_i3, I created a new LIT according to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2501 and added this new ID to every manuscript record I became aware of.

For ms_i4.2, I was not certain, but I think it is probable, that it might be LIT6865SalGabreel here. The incipit in the manuscript according to Strelcyn is ሰላም፡ ለከ፡ ገብርኤል፡ ብስራታዊ። The incipit in the work record has ሰላም፡ ገብርኤል፡ ብስራታዊ።

For a prayer to the lord in a3, I have just opened https://github.com/BetaMasaheft/Documentation/issues/2502 a minute ago and I would like to read your opinions on that, before I create a new LIT ID.

For a6, I did not know, if I should use LIT3308Nicene or any other LIT for the Creed, because I did not know, if this was the conventional EOTC creed (what is very likely) or any alternative creed, since Strelcyn did not quote it or refer to it in a specific way.

I struggled with the question, whether ms_i2, ms_i3, ms_i4 and ms_i5 should be encoded as part of ms_i1, as it was defined as supplication, doxology or explicit in other Psalter manuscript records (e.g. ESamm013 or BAVet105). I preferred to define them as their own msItems, because I find, they are individual, even though they are embedded or directly following the Psalter.