Create BLorient13269
I created this record according to S. Strelcyns catalogue. I not pretty sure, whether LIT6563HassabHeg is the correct ID for Ḥasāba ḥǝgg (Calendar of movable feasts) in ms_i1. I think it is the ID for all calender computation for movable feast no matter this calculation starts and ends in differnt years and uses different cycles. For ms_i2, I was even less certain. Strelcyn refers to it simply as "Kidan" as it was the case in BLorient8813. But here, there are no incipit or any other information, so I can do nothing more but speculate whether it is LIT2260salota again or LIT1938Mashaf or any other work with this name. I think, it might be better to leave it without markup.