Create BLorient13273 according to Strelcyn
I created a record according to S. Strelcyns catalogue. However, I was not able to identify ms_i6 (https://github.com/BetaMasaheft/Documentation/issues/2508), ms_i8 (https://github.com/BetaMasaheft/Documentation/issues/2509) and a2. For the first and the second, I suggest to create new LIT IDs as argued in the issues. The magical prayer in a2 is probably also a candidate for a new LIT ID, but I did not find any other witness. I checked manuscripts containing LIT4302Prayer and LIT5876AsmatPrayer, but to my point of view, this piece is to much different to identify it as one of these.