Create BLorient8823 (#2271)
* Create BLorient8823

I have made an record for BLorient8823 according to the catalogue of S. Strelcyn. I was not able to identify the works listed in Strelcyn's catalogue as "Three Compositions for Moral Edification", and I do not know if they have any circulation apart from this manuscript. I have therefore kept them as described by Strelcyn. I discovered in Emmel's article that images are in principle available online, but as long as the BL server is down for any length of time, I cannot check and investigate more than what is given as incipits in Strelcyn.

* Update BLorient8823

I rearranged provenance, because I found, that acquisition fits better for some of the information.

* Update BLorient8823

I updated the record according to Eugenias and Dorotheas requests.

* Update BLorient8823

I updated the binding in the record according to Eugenias request.

* Update BLorient8823

I updated the msContents according to Eugenias request.

* Update BLorient8823

I updated the decoNote in binding desc concerning the leaf string markers according to Eugenias request.

* Update BLorient8823

I updated the record according to Dorotheas request in https://github.com/BetaMasaheft/Documentation/issues/2482.

* Update BLorient8823

I updated the record according to Eugenias request.