Create BLorient13264 (#2283)
* Create BLorient13264

I created a record for BLorient13264. I could not identify the prayer in ms_i7. I checked Leanders edition of the Prayers of Simeon Stylites and the unidentified prayer is not part of this edition.

I made a minor change also in BAVet57, because I have discovered, that the prayer of Ephraim encoded there is to be identified with LIT4239Prayer as well.

* Update BLorient13264

I updated BLorient13264 according to Eugenias requests.

* Updated BLorient13264 according to Dorotheas requests