Create BLorient11294
For the identification of ms_i2 cf.
https://github.com/BetaMasaheft/Documentation/issues/2549.

I was in difficulties to understand, whether what Strelcyn described in BLorient11294 was a composite manuscript or not. He stated, that the second part is "for" another person as the first part. However ms_i2 start on the same folio as the previous textual unit ends. For this reason, I do not think, that these are codicological units but textual units, that have been sponsored by different patrons.