Create BLorient11718A
This manuscript has no uniform pagination or foliation. The first part is a manuscript in the proper sense written in ink (important addition in pencil) and numbered page 1 to 82 plus unnumbered and unwritten pages, that are most likely in the end (except for one after p. 1). The other part consists of the twice typewritten translation of the aforementioned text. I did not dare to transform the foliation of the latter according to the pagination of the former or vice versa. I think, two systems of pagination / foliation will also not work in one record. For this reason, I decided not to provide any loci for p2, but to explain the setup explicitly.

I was also thinking, whether twice the same text in p2 deserves to be defined as p2 and p3. However, Strelcyn did not give any details for the latter copy except for what I noticed in the note.